refactor(config)!: rename all references of plugin-related 'options' to 'config'
Conventionally, in Lua, variables named `opts` are tables that contains
options that you pass to a function. In Neovim, 'config' is what we
would call the configuration of the plugin, or of Neovim itself.

Following these conventions, `opts` is the parameter of the setup
function, while 'config' is the term used for the configuration itself.