fix: duplicated boilerplate on "move type to" code action
I don't really understand why this bug happens, but it must be some type
of race condition. The "move type to SomeFile.cs" code action appears to
trigger the same Neovim autocommand event that boilersharp uses to
detect new C# files, but for some reason the plugin determines
that the file is empty, even though the LSP has written (or is writting)
to it.

This fix defers the execution of the plugin to after the language server
moves all the code to the new file, allowing boilersharp to see that the
file is not empty. The bug can still happen on big classes (more than
500 lines of code).

Fixes #1