fix: ensure that boilerplate gets written to the correct buffer
By the time `write_boilerplate()` is called, `vim.schedule()` might have
given Neovim enough time to exit the current buffer, causing a
discrepancy between the buffer that triggered the event and the buffer
that `write_boilerplate()` is writting to.

By getting the buffer before we call `vim.schedule()` we ensure that the
buffer that triggered the event is the same that boilerplate gets
written to.